refactor: move the dispatch walk into src/driver.rs, unchanged
About 1,370 lines -- the bounded call-graph walk behind
`reachable_from_dispatch`, the directional path recipe built on it, and their
fifteen engine-free tests -- out of an eight-thousand-line module that is
otherwise the MCP tool surface. `server.rs` drops to 6,770 lines.

The point is what comes next rather than tidiness. The walk is about to be
reworked onto dbgscope's typed instructions, given a deadline and cancellation,
and taught to report module+RVA coordinates, and it is about to gain three more
analyses built on the same traversal. Doing that in place would produce a diff
in which a relocation and a rewrite are indistinguishable, so the move lands
first and alone: **no behaviour change**, 703 tests before and 703 after.

Two things it did not move, on purpose. `parse_eval`, `module_pattern` and
`matches_module_pattern` stay in `server.rs` -- they were interleaved with the
walk under one section header but belong to the module filter, not to this.
And `parse_windbg_addr` moved *with* the walk despite `walk.rs` using it too,
because it is the walk's own parser and one definition in the module that owns
it beats a shared utility in the module that no longer does; the two callers
outside now say `crate::driver::`.
